refactor(k8s): setup lifecycle into a separated function
We use a PostStart hook to check nginx syntax. As we will accept custom
Lifecycles from the user, we must keep checking the nginx syntax, even if
the user sets a PostStart hook with a custom Exec handler. Extracting
this logic to a function will help to do it easier.