feat(k8s): append user lifecycle to default lifecycle
We check the nginx syntax in a PostStart handler with an Exec action.
Because of it, if the user defines this kind of handler and action we
cannot just override de default lifecycle syntax not being checked. So
we need to append the user command to the command that checks the syntax.